moved the naming of the Demo Web Client to make more sense.  Added some missing details from some of the pom.xml files.
