+++ CLClient/src/main/java/ws/argo/CLClient/commands/SendProbe.java
@@ -1,6 +1,5 @@
 package ws.argo.CLClient.commands;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
