+++ DemoWebClient/ArgoWebClient/pom.xml
@@ -7,7 +7,7 @@
     <artifactId>DemoWebClient</artifactId>
     <version>0.3.0-SNAPSHOT</version>
   </parent>
-  <groupId>ws.argo.Browser</groupId>
+  <groupId>ws.argo.DemoWebClient</groupId>
   <artifactId>ArgoWebClient</artifactId>
   <version>0.3.0-SNAPSHOT</version>
   <packaging>war</packaging>
