+++ DemoWebClient/ArgoWebClient/src/main/java/ws/argo/DemoWebClient/Browser/ProbeRespondToAddress.java
@@ -1,4 +1,4 @@
-package ws.argo.BrowserWeb;
+package ws.argo.DemoWebClient.Browser;
 
 class ProbeRespondToAddress {
 	public String respondToAddress;
