+++ DemoWebClient/AsynchListener/pom.xml
@@ -8,7 +8,7 @@
 		<artifactId>DemoWebClient</artifactId>
 		<version>0.3.0-SNAPSHOT</version>
 	</parent>
-	<groupId>ws.argo.AsynchListener</groupId>
+	<groupId>ws.argo.DemoWebClient</groupId>
 	<artifactId>AsynchListener</artifactId>
 	<version>0.3.0-SNAPSHOT</version>
 	<packaging>war</packaging>
