+++ ProbeGenerator/pom.xml
@@ -33,4 +33,6 @@
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
+	<name>Probe Generator</name>
+	<description>Provides the capability for Argo clients to create and send discovery probes.  Very small and easy to use.</description>
 </project>
